requirements.txt
decord
diffusers
imageio
einops
gdown
numpy
matplotlib
opencv_python
Pillow
safetensors
torch==2.4.0
torchvision
torchaudio
tqdm
lpips
wandb
accelerate
